Create HTML 5
In a section for a school pattern such as Aptech, we have structured and unstructured codes. The styler follows a certain HTML standard at a certain point in time, The header figure was misinterpreted by the styler who failed to follow the HTML standard. The html5 consist of six different unique tags and follows the default html standard as a styler, deduce the context above.